refactor(nav): extract DragGestureController from ExperimentalControls
Move the merged pan+rotate mouse gesture core into dragGestureController.js: the
letterbox sub-mode comparator (the nav-experimental:modechange LB stream), the
rotation ring indicator, the live-Shift pan<->rotate switch, all three pan
sub-modes (screen / truck / pedestal), both rotation regimes (Map orbit / Street
rotate-in-place), and the WASD-yield rotation-end.

The orchestrator keeps the mouse entry points (_onMouseDown / _onMouseMove /
_onMouseUp) as thin routers: O owns only the passive-input guard, the
window-listener attach/detach, the mid-tween abort, and the gesture-end recovery
call; everything gesture-specific lives behind drag.beginGesture / onMove /
endGesture. The key routers' Shift-sync and WASD-yield seams delegate to
drag.syncDragModeToShift / endRotationForWasd. getCurrentLbMode stays callable on
the instance (the useNavMode hook) as a delegator. ctx gains domElement,
emitModeChange, isInactive, mapPivotBoundsRadius, mapPivotFarAcceptGain,
rotationSpeed, drag; emitLbModeChange retargets to the drag controller.

viewRayGroundPoint — a pure geometry helper shared by the Map-orbit pivot and the
compass plan-view pivot — moves to navMath.js (its natural home) rather than
coupling the compass to the drag controller.

Behaviour-preserving: 302/302 characterization tests green; the
rotationEndForWasd break-it re-points to the drag controller.
